requirements.txt
annotated-types==0.7.0
anyio==4.8.0
apt-clone==0.2.1
apturl==0.5.2
bcrypt==4.3.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
eyeD3==0.8.10
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.6.0
grpcio==1.30.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.5
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
more-itertools==8.10.0
nemo-emblems==5.8.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
onboard==1.4.1
packaging==21.3
PAM==0.4.2
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyelftools==0.27
Pygments==2.19.1
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.11.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-jose==3.4.0
python-magic==0.4.24
python-multipart==0.0.20
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
SecretStorage==3.3.1
setproctitle==1.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
starlette==0.45.3
systemd-python==234
tinycss2==1.1.1
tldextract==3.1.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
uvicorn==0.34.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.0.4
webencodings==0.5.1
websockets==15.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zipp==1.0.0
